refactor(blorktools): bone util
